massive refactoring of routing/display for JSON
Pretty much everything was touched to make this happen but the end result probably worth it.  The command_router is now gone and everything is done in one fairly small
page_router() living in controller.php.  The other major change is that every view is now in a function that takes some data returns the HTML.  The decoupling of the data and display allows from JSON dumps (web-services FTW).

routing parameters now exist as follows:
mode=<html (default),json,xml>
type=<overview (default),  hosts,services,hostgroups,servicegroups,hostdetail,servicedetail,object>
*state_filter=UP,DOWN,WARNING,UNREACHABLE,UNKNOWN,CRITICAL
*name_filter=<string>
*objtype_filter=hosts_objs,services_objs,hostgroups_objs,servicegroups_objs,timeperiods,contacts,contactgroups,commands

Please note the filters can be chained so you can find all of the down services with SSH in the name.  That's kinda handy.  Simply set the mode to JSON and you get a JSON dump.  XML support is not rolled back in at this time.